product_pricelist_direct_print: Add hooks
