+++ product_pricelist_direct_print/wizards/product_pricelist_print.py
@@ -93,6 +93,7 @@ class ProductPricelistPrint(models.TransientModel):
 
     @api.depends_context("product")
     def _compute_product_price(self):
+        # TODO drop this computed field
         product = self.env.context["product"]
         price = self.get_pricelist_to_print()._get_product_price(
             product, 1, date=self.date
@@ -108,6 +109,20 @@ class ProductPricelistPrint(models.TransientModel):
             )
         else:
             self.product_price = float_round(price, precision)
+
+    def get_price_for_pricelist(self, pricelist, product):
+        # TODO enable this instead
+        # pricelist.ensure_one()
+        # product.ensure_one()
+        # price = pricelist._get_product_price(
+        # product, 1, date=self.date
+        # )
+        # if self.vat_mode == "vat_excl":
+        # return product.taxes_id.compute_all(price)["total_excluded"]
+        # elif self.vat_mode == "vat_incl":
+        # return product.taxes_id.compute_all(price)["total_included"]
+        # return price
+        return self.with_context(product=product).product_price
 
     @api.depends("partner_ids")
     def _compute_partner_count(self):
@@ -192,13 +207,17 @@ class ProductPricelistPrint(models.TransientModel):
             [(field.name, field.display_name) for field in fields], key=lambda f: f[1]
         )
 
-    def print_report(self):
-        if not (
+    def _allowed_to_print_report(self):
+        self.ensure_one()
+        return (
             self.pricelist_id
             or self.partner_count
             or self.show_standard_price
             or self.show_sale_price
-        ):
+        )
+
+    def print_report(self):
+        if not self._allowed_to_print_report():
             raise ValidationError(
                 _(
                     "You must set price list or any customer "
